+++ go/pkg/db/migrations_integration_test.go
@@ -28,6 +28,8 @@ func TestLiveMigrationsInstallCurrentSchemaInvariants(t *testing.T) {
 		"repo_event_chain_heads",
 		"auto_finalize_circuit_breakers",
 		"job_recovery_state",
+		"fanin_freeze_points",
+		"barrier_staged_contributions",
 	} {
 		var exists bool
 		if err := pool.RawPool.QueryRow(ctx, `